+++ index.html
@@ -10,8 +10,6 @@
 </head>
 
 <body>
-  <script src="main.js"></script>
-  <script src="js/functions.js"></script>
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
@@ -233,6 +231,8 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
+
+  <script src="main.js" type="module"></script>
 
 </body>
 </html>
